Skill_Hub/requirements.txt
# Skill Hub Requirements

# CLI framework
click>=8.1.0

# Configuration
pyyaml>=6.0

# Machine Learning (optional, for better similarity detection)
scikit-learn>=1.3.0

# HTTP client (for GitHub API)
httpx>=0.24.0

# Async support
asyncio>=3.4.3
